+++ CHANGELOG.md
@@ -7,3 +7,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.0 - 2025-01-30]
 ### Added
 - First version of the project
+- GitHub CI
+- REST services: POST and DELETE beam_app